move torch helper function of padding in unets to torch_helpers.
